pyproject.toml
[project]
name = "windows-automation-agent"
version = "0.1.0"
description = "AI-powered Windows GUI automation agent using LangGraph and pywinauto"
authors = [
    {name = "Devansh", email = "devansh.arctikcircle@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph>=1.0.3,<2.0.0",
    "langchain>=1.0.7,<2.0.0",
    "langchain-openai>=1.0.3,<2.0.0",
    "langchain-core>=1.0.5,<2.0.0",
    "pydantic>=2.12.4,<3.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "typer>=0.12.0,<1.0.0",
    "rich>=13.7.0,<14.0.0",
    "pillow>=12.0.0,<13.0.0",
]

[project.optional-dependencies]
windows = [
    "pywinauto>=0.6.8,<0.7.0",
]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "ruff>=0.4.0,<1.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
